factor out LoadRemover
